removal logic - MarkForRemovalButtonPress
Vinyl can be marked for remove in any state. (I thought that if I discover an error on a vinyl when it is already borrowed/B&r or reserved but valid, and this vinyl can no longer be used after it becomes valid again)
But the actual change of state to "available for remove" is only done when the vinyl is no longer borrowed or reserved.